+++ src/dolphin/os/OSArena.c
@@ -1,7 +1,7 @@
 #include <dolphin/os/OSArena.h>
 
-#define ROUND(n, a) (((u32)(n) + (a)-1) & ~((a)-1))
-#define TRUNC(n, a) (((u32)(n)) & ~((a)-1))
+#define ROUND(n, a) (((uintptr_t)(n) + (a)-1) & ~((a)-1))
+#define TRUNC(n, a) (((uintptr_t)(n)) & ~((a)-1))
 
 void *__OSArenaHi;
 void *__OSArenaLo = (void *)-1;
